+++ Opt/UtilityExtensions/IEnumerableExt.cs
@@ -104,5 +104,14 @@ namespace ComradeVanti.CSharpTools
         /// <returns>The present values</returns>
         public static IEnumerable<T> FilterSome<T>(this IEnumerable<IOpt<T>> opts) =>
             opts.Choose(it => it);
+
+        /// <summary>
+        ///     Attempts to get the maximum element from the collection
+        /// </summary>
+        /// <param name="items">The collection</param>
+        /// <typeparam name="T">The type of the contained value</typeparam>
+        /// <returns>The maximum. None if the collection is empty</returns>
+        public static IOpt<T> TryMax<T>(this IEnumerable<T> items) =>
+            Opt.FromOp(items.Max);
     }
 }